[omm] Flask app skeleton in dev container + docs
Initially aiming for a turnkey developer experience before implementing
any real functionality.

Setup a bare-bones Flask app skeleton with a couple of toy endpoints
to get things started.

Configure the dev container and document the onboarding process in
the CONTRIBUTING.md file.
