+++ open-media-match/src/OpenMediaMatch/__init__.py
@@ -1,1 +1,30 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates.
+
+import flask
+
+
+def create_app():
+    """
+    Create and configure the Flask app
+    """
+    app = flask.Flask(__name__, instance_relative_config=True)
+
+    @app.route("/")
+    def ping():
+        """
+        Liveness/readiness check endpoint for your favourite Layer 7 load balancer
+        """
+        return "I-AM-ALIVE"
+
+    @app.route("/hash/image")
+    def hash_image():
+        """
+        Hash an image
+        """
+        image_url = flask.request.args.get('url')
+        if image_url is None:
+            flask.abort(400)
+
+        return f"If this was implemented you'd be seeing the PDQ hash of the image at {image_url} here"
+
+    return app
